Preparatory refactor: move the statement CSV parser into src/lib
Behaviour identical. `parseCsv`, `normalizeDate` and `toBankRows` move out of
`/manager/reconciliation` and into `src/lib/bank-statement-csv.ts` verbatim, so
they can be imported and unit tested directly rather than through a rendered
tree. That is the seam `validation.ts` and `submit-resilience.ts` already
establish; nothing about the parsing changes here.

No tests yet: that is the next commit, so the pinning tests can be seen to pass
against the code exactly as it has been running.

Refs #65
